use of netsh to get network name
Sometimes the connection event didn't give the network name, causing the script to activate the tunnel all the time, even if the network was secure. By using netsh, you get the network SSID all the time. However, you lose the ability to activate or deactivate the tunnel with wired connections.